requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
strawberry-graphql==0.246.1
httpx==0.27.0
pyjwt==2.9.0
pydantic==2.9.1
